Fixed the thing where changing the active sequence while in notes setup mode wouldn't update the screen.
Improved the control code recording. You can choose whether to record/play it back, and it sends correct CCs while playing live, as well as in the recorded sequence.

I took out the channel aftertouch, so now it just sends poly aftertouch, and the selected CC.

Added tooltip when hovering over the notes in the VirtualLpp debug ui.