User mode.
Programmable sliders or checkboxes that can send any CC on any channel.

Go into setup. The first checkbox is whether that row should be checkboxes (on)
or sliders (off). The next checkbox is whether the slider should be bipolar (on)
or unipolar (off). The number after that is the midi channel.

Hit shift and you'll see 7-pad numbers which represent the CC sent.

For checkbox rows, each checkbox sends a CC one greater than the last, starting
with the configured CC.