External midi clock.
Sequencer mode settings have been moved around and some new checkboxes are
added. One for enabling external midi clock, and another for selecting the
port to accept clock on.

With external midi clock enabled, the sequence will only tick when receiving
midi clock messages.

That's not entirely true if you're using swing, in which case some steps will
happen in between midi clock ticks, but if you set swing to 0, then the timing
should accurately follow your external clock. It should still accurately follow
it with swing as well, just it might be not quite so dead on for the swung steps.