Big fix to sequencer mode.
Now, when you change your global scale, notes that are not included in the
scale are still visible in the sequencer mode, rounded down to the next lowest
note that is in the scale.

Same with notes that are in between steps. When you are zoomed out, you can
still see and edit the notes that are in between steps so you don't have to go
hunting for them.

shift+pad in sequencer mode now not only creates slide notes, but it creates
multiple tied together notes if you are zoomed out, or deletes multiple
tied together notes, so you can edit long held notes more quickly without any
zooming in and out.